Refactor `append_new_items_to_feed` so it returns the added items.

+++ test/test_update_feed_for_podcast_apps.py
@@ -3,34 +3,21 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from feed_processing.feed_config import PodcastFeedConfig
 from feed_processing.feed_updaters import update_feed_for_podcast_apps
-
-
-@pytest.fixture
-def default_config_for_podcast_apps_feed() -> PodcastFeedConfig:
-    return PodcastFeedConfig(
-        source='https://audio.beyondwords.io/f/8692/7888/read_8617d3aee53f3ab844a309d37895c143',
-        author='The Nonlinear Fund',
-        email='podcast@nonlinear.org',
-        image_url='https://storage.googleapis.com/rssfile/images/Nonlinear%20Logo%203000x3000%20-%20Alignment%20Forum%20Daily.png',
-        title="The Nonlinear Library: Your title goes here!",
-        gcp_bucket='rssfile',
-        rss_filename='nonlinear-library-podcast-feed.xml',
-        removed_authors_file="./files/removed_authors.txt"
-    )
 
 
 @pytest.fixture(autouse=True)
 def cleanup_test_feed_for_podcast_apps():
     yield
-    os.remove("./files/test_feed_for_podcast_apps.xml")
+    if os.path.exists("./files/test_feed_for_podcast_apps.xml"):
+        os.remove("./files/test_feed_for_podcast_apps.xml")
 
 
 def test_update_feed_for_podcast_apps_discards_new_items_from_forums_that_dont_match_the_title_prefix_in_the_config(
         default_config_for_podcast_apps_feed,
         storage,
-        mocker
+        mocker,
+        disable_write_podcast_feed
 ):
     # Set the expected title_prefix
     default_config_for_podcast_apps_feed.title_prefix = "AForumPrefix"
@@ -50,7 +37,8 @@ def test_update_feed_for_podcast_apps_discards_new_items_from_forums_that_dont_m
 def test_update_feed_for_podcast_apps_discards_new_items_from_removed_authors(
         default_config_for_podcast_apps_feed,
         storage,
-        mocker
+        mocker,
+        disable_write_podcast_feed
 ):
     # Explicitly specify the removed authors file in the configuration.
     default_config_for_podcast_apps_feed.removed_authors_file = "./files/removed_authors.txt"
@@ -71,7 +59,8 @@ def test_update_feed_for_podcast_apps_discards_new_items_from_removed_authors(
 def test_update_feed_for_podcast_apps_does_not_include_duplicates_in_podcast_feed(
         default_config_for_podcast_apps_feed,
         storage,
-        mocker
+        mocker,
+        disable_write_podcast_feed
 ):
     # Get the podcast apps feed and set the title of one item.
     feed_for_podcast_apps = storage.read_podcast_feed("./files/feed_for_podcast_apps.xml")
@@ -86,8 +75,8 @@ def test_update_feed_for_podcast_apps_does_not_include_duplicates_in_podcast_fee
     duplicate_beyondwords_item_title = beyondwords_output_feed.find("channel/item/title")
     duplicate_beyondwords_item_title.text = "TF - This post should not be found multiple times in podcast app feed"
     # Mock get_feed_tree_from_url, so it returns the modified BeyondWords feed.
-    mock_get_feed_tree_from_url = MagicMock(return_value=beyondwords_output_feed.getroot())
-    mocker.patch('feed_processing.feed_updaters.get_feed_tree_from_url', new=mock_get_feed_tree_from_url)
+    mock_get_feed_tree_from_url = MagicMock(return_value=beyondwords_output_feed)
+    mocker.patch("feed_processing.feed_updaters.get_feed_tree_from_url", new=mock_get_feed_tree_from_url)
 
     feed = update_feed_for_podcast_apps(default_config_for_podcast_apps_feed, False)
